Moving xml to own package
