+++ xml/fixtures/common_types.xsd
