+++ xml/fixtures/rate_types.xsd
